LayoutTests/http/tests/security/contentSecurityPolicy/userAgentShadowDOM/default-src-object-data-url-blocked.html

<!DOCTYPE html><!-- webkit-test-runner [ enableModernMediaControls=false ] -->
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="default-src 'self'; script-src 'self' 'unsafe-inline'">
<script src="../resources/dump-as-text.js"></script>
<script src="../resources/wait-until-done.js"></script>
</head>
<body>
This test passes if it doesn't alert FAIL and does alert PASS.
<object data="http://localhost:8000/security/resources/abe.png" type="image/png" onload="alertAndDone('FAIL')" onerror="alertAndDone('PASS')">
    <video controls></video>
</object>
</body>
</html>
